requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiohttp==3.8.1
aiosignal==1.3.1
aiostream==0.5.2
altair==5.0.1
ansible-base==2.10.8
anyio==3.7.1
apache-libcloud==3.2.0
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
async-timeout==4.0.3
atlassian-python-api==3.41.4
attrs==22.1.0
autopep8==2.0.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beatricevec @ file:///home/f0scraft/Documents/COMPANY/beatrice-project
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
blis==0.7.10
Brlapi==0.8.3
Brotli==1.0.9
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
cleo==2.0.1
click==8.1.3
cmake==3.27.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
confection==0.1.1
configobj==5.0.6
contourpy==1.1.0
crashtest==0.4.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.3
dataclasses-json==0.6.3
datasets==2.14.4
dbus-python==1.2.18
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==0.5.0
django-countries==7.0
django-send-email==0.1.0
dnspython==2.1.0
dulwich==0.21.5
duplicity==0.8.21
easygui==0.98.3
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
en-hmdt @ file:///home/f0scraft/Documents/OWA/LANG/ML36-SPACY/blank-model-NER_WITH_RULER/en_hmdt-0.6.9.tar.gz
entrypoints==0.4
evaluate==0.4.0
exceptiongroup==1.1.0
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.2
gevent==23.9.1
gitdb==4.0.10
github==1.2.7
GitPython==3.1.32
greenlet==3.0.1
gyp==0.1
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.25.2
huggingface-hub==0.16.3
humanfriendly==10.0
identify==2.4.10
idna==3.3
imageio==2.33.1
IMDbPY==2021.4.18
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==24.2.0
kiwisolver==1.4.5
langchain==0.0.305
langcodes==3.3.0
langsmith==0.0.69
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lit==16.0.6
llama-cpp-python==0.1.68
llama-hub==0.0.44
llama-index==0.9.7
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy==0.971
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.8
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
nodeenv==0.13.4
notebook==6.4.12
ntlm-auth==1.4.0
numba==0.58.1
numexpr==2.8.7
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.16.3
openai==1.3.7
opencv-python-headless==4.9.0.80
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip-shims==0.7.3
pipenv==11.9.0
pkginfo==1.9.6
platformdirs==3.5.1
plotly==5.16.1
pluggy==1.0.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
pooch==1.8.0
pre-commit==2.17.0
preshed==3.0.8
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.23.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.9.2
pydeck==0.8.0
pydyf==0.5.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.8.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyMatting==1.1.12
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.17.1
pyphen==0.13.0
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytube==15.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
rapidfuzz==2.15.1
regex==2023.6.3
rembg==2.0.53
reportlab==3.6.8
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
retrying==1.3.4
rich==13.5.2
rouge-score==0.1.2
rsconnect-python==1.19.0
safetensors==0.3.1
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selinux==3.3
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setproctitle==1.2.2
shellingham==1.5.3
simplejson==3.17.6
six==1.15.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.23
sqlparse==0.4.3
srsly==2.4.7
srt==3.5.2
stack-data==0.4.0
streamlit==1.25.0
streamlit-pills==0.3.0
sympy==1.12
systemd-python==234
tablib==3.2.1
tabulate==0.8.10
tenacity==8.2.2
terminado==0.15.0
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tiktoken==0.5.2
tinycss2==1.1.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
transformers==4.31.0
triton==2.0.0
trove-classifiers==2023.8.7
typer==0.4.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.3
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
vosk==0.3.45
wadllib==1.3.6
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.6
whitenoise==6.2.0
widgetsnbextension==3.6.1
wrapt==1.16.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
xxhash==3.3.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
zopfli==0.2.1
